requirements.txt
requests>=2.28
jinja2>=3.1
python-dotenv>=1.0